address PR comments
